added vacancies
